Feat/improve question deleting system

+++ src/components/QuestionsList.tsx
@@ -5,6 +5,7 @@ import { Company, Position, Question } from "@/types";
 import {
   ColumnDef,
   ColumnFiltersState,
+  PaginationState,
   SortingState,
   VisibilityState,
   flexRender,
@@ -14,7 +15,7 @@ import {
   getSortedRowModel,
   useReactTable,
 } from "@tanstack/react-table";
-import { ArrowUpDown } from "lucide-react";
+import { ArrowUpDown, ListPlus, Trash } from "lucide-react";
 import {
   Table,
   TableBody,
@@ -27,11 +28,10 @@ import { toast } from "react-toastify";
 
 import { Button } from "@/components/ui/button";
 import { Input } from "@headlessui/react";
-import { FaLeftLong, FaRightLong } from "react-icons/fa6";
 import { QuestionDialog } from "./QuestionDialog";
-import { FaTrash } from "react-icons/fa";
 import { ConfirmationDialog } from "./ConfirmationDialog";
 import { QuestionsFilter } from "./QuestionsFilter";
+import { DialogDescription, DialogTitle } from "./ui/dialog";
 
 export default function QuestionsList({
   company,
@@ -45,14 +45,37 @@ export default function QuestionsList({
   const [columnVisibility, setColumnVisibility] = useState<VisibilityState>({});
   const [rowSelection, setRowSelection] = useState({});
   const [_itemsPerPage, _setItemsPerPage] = useState<number>(10);
-  const [currentPage, setCurrentPage] = useState<number>(0);
-  const [displayedQuestions, setDisplayedQuestions] = useState<Question[]>([]);
   const [questions, setQuestions] = useState<Question[]>([]);
   const [validQuestions, setValidQuestions] = useState<Question[]>([]);
-  const [totalPages, setTotalPages] = useState<number>(0);
 
   const [fetchingQuestions, setFetchingQuestions] = useState(true);
   const [selectedPositions, setSelectedPositions] = useState<number[]>([]);
+
+  const [pagination, setPagination] = useState<PaginationState>({
+    pageIndex: 0,
+    pageSize: 10,
+  });
+
+  const filterQuestions = (positions: any) => {
+    console.log("filtering questions");
+    const filteredQuestions = questions.reduce((acc: any, curr: any) => {
+      curr.positions.map((position: any) => {
+        if (positions.includes(position.position_id)) {
+          if (!acc.includes(curr)) {
+            acc.push(curr);
+          }
+        }
+      });
+      return acc;
+    }, []);
+
+    if (filteredQuestions.length === 0) {
+      setValidQuestions(questions);
+      return;
+    }
+
+    setValidQuestions(filteredQuestions);
+  };
 
   const fetchQuestions = async () => {
     console.log("fetching questions");
@@ -78,12 +101,9 @@ export default function QuestionsList({
       console.log(sortedQuestions);
       setQuestions(sortedQuestions);
       setValidQuestions(sortedQuestions);
-      setTotalPages(Math.ceil(sortedQuestions.length / 10));
-      setDisplayedQuestions(sortedQuestions.slice(0, 10));
       return;
     }
     setQuestions([]);
-    setDisplayedQuestions([]);
   };
 
   const handlePositionClick = (positionId: number) => {
@@ -94,69 +114,17 @@ export default function QuestionsList({
       newPositions = [...selectedPositions, positionId];
     }
 
-    const filteredQuestions = questions.reduce((acc: any, curr: any) => {
-      curr.positions.map((position: any) => {
-        if (newPositions.includes(position.position_id)) {
-          if (!acc.includes(curr)) {
-            acc.push(curr);
-          }
-        }
-      });
-      return acc;
-    }, []);
-
-    console.log({ filteredQuestions });
     setSelectedPositions(newPositions);
 
-    if (filteredQuestions.length === 0) {
-      setValidQuestions(questions);
-      setTotalPages(Math.ceil(questions.length / 10));
-      setCurrentPage(0);
-      setDisplayedQuestions(questions.slice(0, 10));
-      return;
-    }
-
-    console.log({ filteredQuestions });
-    console.log({ sliced: filteredQuestions.slice(0, 10) });
-    setValidQuestions(filteredQuestions);
-    setTotalPages(Math.ceil(filteredQuestions.length / 10));
-    setCurrentPage(0);
-    setDisplayedQuestions(filteredQuestions.slice(0, 10));
+    filterQuestions(newPositions);
   };
 
   useEffect(() => {
     if (fetchingQuestions) {
       fetchQuestions();
       setFetchingQuestions(false);
-      setCurrentPage(0);
     }
   }, [company, fetchingQuestions]);
-
-  const handlePageDown = () => {
-    if (currentPage === 0) {
-      return;
-    }
-    const newPage = currentPage - 1;
-    const start = newPage * 10;
-    const end = start + 9;
-    console.log({ start, end });
-    setDisplayedQuestions(validQuestions.slice(start, end));
-    console.log(questions.slice(start, end));
-    setCurrentPage(newPage);
-  };
-
-  const handlePageUp = () => {
-    if (currentPage === totalPages - 1) {
-      return;
-    }
-    const newPage = currentPage + 1;
-    const start = newPage * 10;
-    const end = start + 9;
-    console.log({ start, end });
-    setDisplayedQuestions(validQuestions.slice(start, end));
-    console.log(questions.slice(start, end));
-    setCurrentPage(newPage);
-  };
 
   const columns: ColumnDef<Question>[] = [
     {
@@ -174,16 +142,25 @@ export default function QuestionsList({
         );
       },
       cell: ({ row }) => (
-        <div className=" pl-3 w-full flex flex-col items-start">
-          {row.original.text}
-          <div className="mt-4">
-            <div className="flex flex-row fles-wrap">
-              {(row.original.tags ?? []).length > 0 &&
-                (row.original.tags ?? []).map((tag) => {
+        <div className=" pl-3 w-full flex flex-col items-start justify-center ">
+          <div className="flex flex-row flex-start gap-x-3 items-center ">
+            {row.original.removed && (
+              <div className="bg-red-700 text-white px-1 py-1 text-xs rounded-sm font-bold">
+                Desactivada
+              </div>
+            )}
+            {row.original.text}
+          </div>
+
+          {(row.original.tags ?? []).length > 0 && (
+            <div className="mt-4">
+              <div className="flex flex-row fles-wrap">
+                {(row.original.tags ?? []).map((tag) => {
                   return <Tag key={tag.id} text={tag.tag} />;
                 })}
+              </div>
             </div>
-          </div>
+          )}
         </div>
       ),
     },
@@ -223,16 +200,37 @@ export default function QuestionsList({
             />
 
             <ConfirmationDialog
-              title="Eliminar pregunta"
-              description="¿Seguro que deseas eliminar la pregunta?"
-              confirmText="Eliminar"
+              confirmText={row.original.removed ? "Agregar" : "Eliminar"}
+              mode={row.original.removed ? "add" : "delete"}
               handleSubmit={() => {
-                handleDeleteQuestion(row.original.id, () => {
-                  setFetchingQuestions(true);
-                });
+                //TODO: Que no se vaya a la pagina 0 al hacer fetch de preguntas
+
+                if (row.original.removed) {
+                  handleAddQuestion(row.original.id, () => {
+                    setFetchingQuestions(true);
+                    filterQuestions(selectedPositions);
+                  });
+                  row.original.removed = false;
+                } else {
+                  handleDeleteQuestion(row.original.id, () => {
+                    setFetchingQuestions(true);
+                    filterQuestions(selectedPositions);
+                  });
+                }
               }}
-              triggerText={<FaTrash />}
-            />
+              triggerText={row.original.removed ? <ListPlus /> : <Trash />}
+            >
+              <DialogTitle className="mt-1">
+                {row.original.removed
+                  ? "Agergar pregunta"
+                  : "Eliminar pregunta"}
+              </DialogTitle>
+              <DialogDescription>
+                {row.original.removed
+                  ? "La pregunta se encuentra desactivada. ¿Deseas activarla de nuevo?"
+                  : "Si la pregunta ha sido respondida en alguna evaluación, será desactivada, de lo contrario será eliminada."}
+              </DialogDescription>
+            </ConfirmationDialog>
           </div>
         );
       },
@@ -240,7 +238,7 @@ export default function QuestionsList({
   ];
 
   const table = useReactTable({
-    data: displayedQuestions,
+    data: validQuestions,
     columns,
     onSortingChange: setSorting,
     onColumnFiltersChange: setColumnFilters,
@@ -250,13 +248,15 @@ export default function QuestionsList({
     getFilteredRowModel: getFilteredRowModel(),
     onColumnVisibilityChange: setColumnVisibility,
     onRowSelectionChange: setRowSelection,
-    manualPagination: true,
+
     state: {
       sorting,
       columnFilters,
       columnVisibility,
       rowSelection,
+      pagination,
     },
+    onPaginationChange: setPagination,
     globalFilterFn: (row, _columnId, filterValue) => {
       const question = row.original.text.toLowerCase();
       const searchValue = filterValue.toLowerCase();
@@ -267,6 +267,8 @@ export default function QuestionsList({
 
   return (
     <div className="flex flex-col">
+      {JSON.stringify(table.getState().pagination)}
+      {JSON.stringify(table.getCanNextPage())}
       <QuestionDialog
         company={company}
         setFetchingQuestions={setFetchingQuestions}
@@ -284,45 +286,46 @@ export default function QuestionsList({
             placeholder="Buscar..."
             onChange={(event) => {
               if (event.target.value === "") {
-                setDisplayedQuestions(validQuestions.slice(0, 10));
-                setTotalPages(Math.ceil(validQuestions.length / 10));
-                setCurrentPage(0);
                 return table.setGlobalFilter(event.target.value);
               }
-              const searchedQuestions = validQuestions.filter((question) =>
-                question.text
-                  .toLowerCase()
-                  .includes(event.target.value.toLowerCase())
-              );
-              setDisplayedQuestions(searchedQuestions);
-              setTotalPages(Math.ceil(searchedQuestions.length / 10));
-              setCurrentPage(0);
+
               return table.setGlobalFilter(event.target.value);
             }}
             className="max-w-sm border border-gray-300 rounded-md px-2 py-1"
           />
-          <div className=" text-center">
-            <Button
-              onClick={() => {
-                handlePageDown();
-              }}
-              variant="outline"
-              className="ml-2"
-            >
-              <FaLeftLong />
-            </Button>
-            <Button
-              onClick={() => {
-                handlePageUp();
-              }}
-              variant="outline"
-              className="ml-2"
-            >
-              <FaRightLong />
-            </Button>
-            <p>
-              Página {currentPage + 1} de {totalPages}
-            </p>
+          <div className=" text-center flex flex-col ">
+            {validQuestions.length > 0 && (
+              <div className="flex  flex-col gap-4 lg:gap-0 items-center justify-between space-x-2 py-4 px-5">
+                <div className="flex items-center justify-between mt-4 mb-2 gap-2">
+                  <span className="text-sm text-neutral-600">
+                    Página {pagination.pageIndex + 1} de{" "}
+                    {Math.ceil(validQuestions.length / pagination.pageSize)}
+                    {"."}
+                  </span>
+                  <span className="text-sm text-neutral-600">
+                    Preguntas: {validQuestions.length}
+                  </span>
+                </div>
+                <div className="flex gap-3">
+                  <Button
+                    variant="outline"
+                    size="sm"
+                    onClick={() => table.previousPage()}
+                    disabled={!table.getCanPreviousPage()}
+                  >
+                    Anterior
+                  </Button>
+                  <Button
+                    variant="outline"
+                    size="sm"
+                    onClick={() => table.nextPage()}
+                    disabled={!table.getCanNextPage()}
+                  >
+                    Siguiente
+                  </Button>
+                </div>
+              </div>
+            )}
           </div>
         </div>
 
@@ -396,7 +399,28 @@ function Tag({ text }: { text: string }) {
     </div>
   );
 }
+async function handleAddQuestion(questionId: number, callback: () => void) {
+  const { error } = await supabase
+    .from("questions")
+    .update({ removed: false })
+    .eq("id", questionId);
 
+  if (error) {
+    console.log(error.message);
+    toast.error("Error al agregar la pregunta", {
+      position: "bottom-right",
+      autoClose: 1000,
+    });
+    return;
+  }
+
+  toast.success("Pregunta agregada con éxito", {
+    position: "bottom-right",
+    autoClose: 1000,
+  });
+
+  callback();
+}
 async function handleDeleteQuestion(questionId: number, callback: () => void) {
   //Check if the question has been used in any evaluation
   const { data, error: errorFetchQuestion } = await supabase
@@ -427,22 +451,10 @@ async function handleDeleteQuestion(questionId: number, callback: () => void) {
       return;
     }
 
-    //Add a "deactivated" tag to the question
-    const { error: errorTag } = await supabase.from("question_tags").insert({
-      question_id: questionId,
-      tag: "desactivada",
+    toast.success("Pregunta desactivada correctamente", {
+      position: "bottom-right",
+      autoClose: 1000,
     });
-
-    if (errorTag) {
-      console.log(errorTag.message);
-      toast.error("Error al elminar la pregunta EL03", {
-        position: "bottom-right",
-        autoClose: 1000,
-      });
-      return;
-    }
-
-    //TODO: if this fails, the question should be re added to the database
   } else {
     // delete question from the database
     const response = await supabase
@@ -451,20 +463,18 @@ async function handleDeleteQuestion(questionId: number, callback: () => void) {
       .eq("id", questionId);
     if (response.error) {
       console.log(response.error.message);
-      toast.error("Error al elminar la pregunta EL03", {
+      toast.error("Error al eliminar la pregunta EL03", {
         position: "bottom-right",
         autoClose: 1000,
       });
       return;
     }
 
-    console.log(response);
+    toast.success("Pregunta eliminada correctamente", {
+      position: "bottom-right",
+      autoClose: 1000,
+    });
   }
-
-  toast.success("Pregunta eliminada correctamente", {
-    position: "bottom-right",
-    autoClose: 1000,
-  });
 
   callback();
